move activities into activity package
