+++ app/src/main/AndroidManifest.xml
@@ -14,38 +14,38 @@
         android:theme="@style/Theme.JavaHelpFrontend"
         tools:targetApi="31">
         <activity
-            android:name=".frontend.FrontPage2Activity"
+            android:name=".frontend.activity.FrontPage2Activity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".frontend.ClientRegistrationActivity"
+            android:name=".frontend.activity.ClientRegistrationActivity"
             android:exported="false"
-            android:parentActivityName=".frontend.LoginActivity">
+            android:parentActivityName=".frontend.activity.LoginActivity">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".frontend.ProviderRegistrationActivity"
+            android:name=".frontend.activity.ProviderRegistrationActivity"
             android:exported="false"
-            android:parentActivityName=".frontend.LoginActivity">
+            android:parentActivityName=".frontend.activity.LoginActivity">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".frontend.FrontPageActivity"
+            android:name=".frontend.activity.FrontPageActivity"
             android:exported="false"
-            android:parentActivityName=".frontend.LoginActivity">
+            android:parentActivityName=".frontend.activity.LoginActivity">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".frontend.LoginActivity"
+            android:name=".frontend.activity.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
