+++ app/src/main/java/com/javahelp/frontend/activity/FrontPage2Activity.java
@@ -1,4 +1,4 @@
-package com.javahelp.frontend;
+package com.javahelp.frontend.activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +6,11 @@ import android.os.Bundle;
 
 import com.javahelp.R;
 
-public class FrontPageActivity extends AppCompatActivity {
+public class FrontPage2Activity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_fg);
+        setContentView(R.layout.activity_front_page2);
     }
 }
