+++ app/src/main/res/layout/activity_fg.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:orientation="vertical"
     android:layout_height="match_parent"
-    tools:context=".frontend.FrontPageActivity">
+    tools:context=".frontend.activity.FrontPageActivity">
 
     <TextView
         android:layout_width="match_parent"
