+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="data"
-            type="com.javahelp.frontend.LoginViewModel" />
+            type="com.javahelp.frontend.activity.LoginViewModel" />
 
     </data>
 
@@ -16,7 +16,7 @@
         android:orientation="vertical"
         android:padding="10dp"
         android:background="@drawable/b2"
-        tools:context=".frontend.LoginActivity">
+        tools:context=".frontend.activity.LoginActivity">
 
 
         <TextView
