+++ app/src/main/res/layout/activity_preg.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="pdata"
-            type="com.javahelp.frontend.ProviderRegistrationViewmodel" />
+            type="com.javahelp.frontend.activity.ProviderRegistrationViewmodel" />
 
     </data>
 
@@ -15,7 +15,7 @@
         android:layout_height="match_parent"
         android:orientation="vertical"
         android:background="@color/white"
-        tools:context=".frontend.ProviderRegistrationActivity">
+        tools:context=".frontend.activity.ProviderRegistrationActivity">
 
         <ScrollView
             android:layout_width="match_parent"
